Check if name is used

+++ pages/user/_user.vue
@@ -48,14 +48,20 @@ export default {
           score: this.$route.params.user || this.$store.getters.getCurrentScore || 0
         }
 
-        this.$service.score.add(userData).then(() => {
-          if (process.browser) {
-            window.localStorage.setItem('userName', this.userName)
-          }
-          this.$router.push('/rank')
-        }).catch(error => {
-          if (error.response) {
-            this.error = 'Nie udało się zapisać wyniku, spróbuj jeszcze raz.'
+        this.$service.user.validate(this.userName).then(resp => {
+          if (resp.user_exists) {
+            this.error = 'Podane imię jest już zajęte, podaj inne.'
+          } else {
+            this.$service.score.add(userData).then(() => {
+              if (process.browser) {
+                window.localStorage.setItem('userName', this.userName)
+              }
+              this.$router.push('/rank')
+            }).catch(error => {
+              if (error.response) {
+                this.error = 'Nie udało się zapisać wyniku, spróbuj jeszcze raz.'
+              }
+            })
           }
         })
       } else {
